+++ Sources/PeripheryKit/Project/ProjectDriver.swift
@@ -1,5 +1,6 @@
 import Foundation
 
 protocol ProjectDriver {
+    func build() throws
     func index(graph: SourceGraph) throws
 }